+++ gneiss-mqtt/src/features/gneiss_tokio/mod.rs
@@ -1105,9 +1105,6 @@ impl Drop for TokioClientEventWaiter {
 
 #[cfg(all(test, feature = "testing"))]
 pub(crate) mod testing {
-    use assert_matches::assert_matches;
-    use std::future::Future;
-    use std::pin::Pin;
     use std::time::Duration;
     use crate::client::waiter::*;
     use crate::config::*;
@@ -1116,64 +1113,9 @@ pub(crate) mod testing {
     use super::*;
     use crate::testing::integration::*;
 
-    fn create_client_builder_internal(connect_options: ConnectOptions, _tls_usage: TlsUsage, ws_config: WebsocketUsage, proxy_config: ProxyUsage, tls_endpoint: TlsUsage, ws_endpoint: WebsocketUsage) -> GenericClientBuilder {
-        let client_config = MqttClientOptionsBuilder::new()
-            .with_connect_timeout(Duration::from_secs(5))
-            .with_offline_queue_policy(OfflineQueuePolicy::PreserveAll)
-            .build();
-
-        let endpoint = get_broker_endpoint(tls_endpoint, ws_endpoint);
-        let port = get_broker_port(tls_endpoint, ws_endpoint);
-
-        let mut builder = GenericClientBuilder::new(&endpoint, port);
-        builder.with_connect_options(connect_options);
-        builder.with_client_options(client_config);
-
-        #[cfg(feature = "tokio-rustls")]
-        if _tls_usage == TlsUsage::Rustls {
-            let mut tls_options_builder = TlsOptionsBuilder::new();
-            tls_options_builder.with_verify_peer(false);
-            tls_options_builder.with_root_ca_from_path(&get_ca_path()).unwrap();
-
-            builder.with_tls_options(tls_options_builder.build_rustls().unwrap());
-        }
-
-        #[cfg(feature = "tokio-native-tls")]
-        if _tls_usage == TlsUsage::Nativetls {
-            let mut tls_options_builder = TlsOptionsBuilder::new();
-            tls_options_builder.with_verify_peer(false);
-            tls_options_builder.with_root_ca_from_path(&get_ca_path()).unwrap();
-
-            builder.with_tls_options(tls_options_builder.build_native_tls().unwrap());
-        }
-
-        #[cfg(feature="tokio-websockets")]
-        if ws_config != WebsocketUsage::None {
-            let websocket_options = WebsocketOptionsBuilder::new().build();
-            builder.with_websocket_options(websocket_options);
-        }
-
-        if proxy_config != ProxyUsage::None {
-            let proxy_endpoint = get_proxy_endpoint();
-            let proxy_port = get_proxy_port();
-            let proxy_options = HttpProxyOptionsBuilder::new(&proxy_endpoint, proxy_port).build();
-            builder.with_http_proxy_options(proxy_options);
-        }
-
-        builder
+    fn build_tokio_client(builder: GenericClientBuilder) -> AsyncGneissClient {
+        builder.build_tokio(&tokio::runtime::Handle::current()).unwrap()
     }
-
-    fn create_good_client_builder(tls: TlsUsage, ws: WebsocketUsage, proxy: ProxyUsage) -> GenericClientBuilder {
-        let connect_options = ConnectOptionsBuilder::new()
-            .with_rejoin_session_policy(RejoinSessionPolicy::PostSuccess)
-            .with_session_expiry_interval_seconds(3600)
-            .build();
-
-        create_client_builder_internal(connect_options, tls, ws, proxy, tls, ws)
-    }
-
-    type AsyncTestFactoryReturnType = Pin<Box<dyn Future<Output = MqttResult<()>> + Send>>;
-    type AsyncTestFactory = Box<dyn Fn(GenericClientBuilder) -> AsyncTestFactoryReturnType + Send + Sync>;
 
     fn do_good_client_test(tls: TlsUsage, ws: WebsocketUsage, proxy: ProxyUsage, test_factory: AsyncTestFactory) {
         let runtime = tokio::runtime::Runtime::new().unwrap();
@@ -1189,52 +1131,11 @@ pub(crate) mod testing {
         runtime.block_on(test_future).unwrap();
     }
 
-    async fn start_client(client: &AsyncGneissClient) -> MqttResult<()> {
-        let mut connection_attempt_waiter = TokioClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionAttempt);
-        let mut connection_success_waiter = TokioClientEventWaiter::new_single(client.clone(), ClientEventType::ConnectionSuccess);
-
-        client.start(None)?;
-
-        connection_attempt_waiter.wait().await?;
-        let connection_success_events = connection_success_waiter.wait().await?;
-        assert_eq!(1, connection_success_events.len());
-        let connection_success_event = &connection_success_events[0].event;
-        assert_matches!(**connection_success_event, ClientEvent::ConnectionSuccess(_));
-        if let ClientEvent::ConnectionSuccess(success_event) = &**connection_success_event {
-            assert_eq!(ConnectReasonCode::Success, success_event.connack.reason_code);
-        } else {
-            panic!("impossible");
-        }
-
-        Ok(())
-    }
-
-    async fn stop_client(client: &AsyncGneissClient) -> MqttResult<()> {
-        let mut disconnection_waiter = TokioClientEventWaiter::new_single(client.clone(), ClientEventType::Disconnection);
-        let mut stopped_waiter = TokioClientEventWaiter::new_single(client.clone(), ClientEventType::Stopped);
-
-        client.stop(None)?;
-
-        let disconnect_events = disconnection_waiter.wait().await?;
-        assert_eq!(1, disconnect_events.len());
-        let disconnect_event = &disconnect_events[0].event;
-        assert_matches!(**disconnect_event, ClientEvent::Disconnection(_));
-        if let ClientEvent::Disconnection(event) = &**disconnect_event {
-            assert_matches!(event.error, MqttError::UserInitiatedDisconnect(_));
-        } else {
-            panic!("impossible");
-        }
-
-        stopped_waiter.wait().await?;
-
-        Ok(())
-    }
-
     async fn tokio_connect_disconnect_test(builder: GenericClientBuilder) -> MqttResult<()> {
         let client = builder.build_tokio(&tokio::runtime::Handle::current()).unwrap();
 
-        start_client(&client).await?;
-        stop_client(&client).await?;
+        start_client(&client, TokioClientEventWaiter::new_single).await?;
+        stop_client(&client, TokioClientEventWaiter::new_single).await?;
 
         Ok(())
     }
@@ -1334,34 +1235,7 @@ pub(crate) mod testing {
     }
 
     async fn tokio_subscribe_unsubscribe_test(builder: GenericClientBuilder) -> MqttResult<()> {
-        let client = builder.build_tokio(&tokio::runtime::Handle::current()).unwrap();
-        start_client(&client).await?;
-
-        let subscribe = SubscribePacket::builder()
-            .with_subscription_simple("hello/world".to_string(), QualityOfService::AtLeastOnce)
-            .build();
-
-        let subscribe_result = client.subscribe(subscribe, None).await;
-        assert!(subscribe_result.is_ok());
-        let suback = subscribe_result.unwrap();
-        assert_eq!(1, suback.reason_codes.len());
-        assert_eq!(SubackReasonCode::GrantedQos1, suback.reason_codes[0]);
-
-        let unsubscribe = UnsubscribePacket::builder()
-            .with_topic_filter("hello/world".to_string())
-            .with_topic_filter("not/subscribed".to_string())
-            .build();
-
-        let unsubscribe_result = client.unsubscribe(unsubscribe, None).await;
-        assert!(unsubscribe_result.is_ok());
-        let unsuback = unsubscribe_result.unwrap();
-        assert_eq!(2, unsuback.reason_codes.len());
-        assert_eq!(UnsubackReasonCode::Success, unsuback.reason_codes[0]);
-        // broker may or may not give us a not subscribed reason code, so don't verify
-
-        stop_client(&client).await?;
-
-        Ok(())
+        async_subscribe_unsubscribe_test(builder.build_tokio(&Handle::current()).unwrap(), TokioClientEventWaiter::new_single).await
     }
 
     #[test]
@@ -1371,70 +1245,9 @@ pub(crate) mod testing {
         }));
     }
 
-    fn verify_successful_publish_result(result: &PublishResponse, qos: QualityOfService) {
-        match result {
-            PublishResponse::Qos0 => {
-                assert_eq!(qos, QualityOfService::AtMostOnce);
-            }
-            PublishResponse::Qos1(puback) => {
-                assert_eq!(qos, QualityOfService::AtLeastOnce);
-                assert_eq!(PubackReasonCode::Success, puback.reason_code);
-            }
-            PublishResponse::Qos2(qos2_result) => {
-                assert_eq!(qos, QualityOfService::ExactlyOnce);
-                assert_matches!(qos2_result, Qos2Response::Pubcomp(_));
-                if let Qos2Response::Pubcomp(pubcomp) = qos2_result {
-                    assert_eq!(PubcompReasonCode::Success, pubcomp.reason_code);
-                }
-            }
-        }
-    }
-
-    fn verify_publish_received(event: &PublishReceivedEvent, expected_topic: &str, expected_qos: QualityOfService, expected_payload: &[u8]) {
-        let publish = &event.publish;
-
-        assert_eq!(expected_qos, publish.qos);
-        assert_eq!(expected_topic, &publish.topic);
-        assert_eq!(expected_payload, publish.payload.as_ref().unwrap().as_slice());
-    }
-
     async fn tokio_subscribe_publish_test(builder: GenericClientBuilder, qos: QualityOfService) -> MqttResult<()> {
         let client = builder.build_tokio(&tokio::runtime::Handle::current()).unwrap();
-        start_client(&client).await?;
-
-        let payload = "derp".as_bytes().to_vec();
-
-        // tests are running in parallel, need a unique topic
-        let uuid = uuid::Uuid::new_v4();
-        let topic = format!("hello/world/{}", uuid.to_string());
-        let subscribe = SubscribePacket::builder()
-            .with_subscription_simple(topic.clone(), QualityOfService::ExactlyOnce)
-            .build();
-
-        let _ = client.subscribe(subscribe, None).await?;
-
-        let mut publish_received_waiter = TokioClientEventWaiter::new_single(client.clone(), ClientEventType::PublishReceived);
-
-        let publish = PublishPacket::builder(topic.clone(), qos)
-            .with_payload(payload.clone())
-            .build();
-
-        let publish_result = client.publish(publish, None).await?;
-        verify_successful_publish_result(&publish_result, qos);
-
-        let publish_received_events = publish_received_waiter.wait().await?;
-        assert_eq!(1, publish_received_events.len());
-        let publish_received_event = &publish_received_events[0].event;
-        assert_matches!(**publish_received_event, ClientEvent::PublishReceived(_));
-        if let ClientEvent::PublishReceived(event) = &**publish_received_event {
-            verify_publish_received(event, &topic, qos, payload.as_slice());
-        } else {
-            panic!("impossible");
-        }
-
-        stop_client(&client).await?;
-
-        Ok(())
+        async_subscribe_publish_test(client, qos, TokioClientEventWaiter::new_single).await
     }
 
     #[test]
@@ -1460,51 +1273,7 @@ pub(crate) mod testing {
 
     // This primarily tests that the will configuration works.  Will functionality is mostly broker-side.
     async fn tokio_will_test(builder: GenericClientBuilder) -> MqttResult<()> {
-        let client = builder.build_tokio(&tokio::runtime::Handle::current()).unwrap();
-        let payload = "Onsecondthought".as_bytes().to_vec();
-
-        // tests are running in parallel, need a unique topic
-        let uuid = uuid::Uuid::new_v4();
-        let will_topic = format!("goodbye/cruel/world/{}", uuid.to_string());
-
-        let will = PublishPacket::builder(will_topic.clone(), QualityOfService::AtLeastOnce)
-            .with_payload(payload.clone())
-            .build();
-
-        let connect_options = ConnectOptionsBuilder::new()
-            .with_rejoin_session_policy(RejoinSessionPolicy::PostSuccess)
-            .with_will(will)
-            .build();
-
-        let will_builder = create_client_builder_internal(connect_options, TlsUsage::None, WebsocketUsage::None, ProxyUsage::None, TlsUsage::None, WebsocketUsage::None);
-        let will_client = will_builder.build_tokio(&tokio::runtime::Handle::current()).unwrap();
-
-        start_client(&client).await?;
-        start_client(&will_client).await?;
-
-        let subscribe = SubscribePacket::builder()
-            .with_subscription_simple(will_topic.clone(), QualityOfService::ExactlyOnce)
-            .build();
-        let _ = client.subscribe(subscribe, None).await?;
-
-        let mut publish_received_waiter = TokioClientEventWaiter::new_single(client.clone(), ClientEventType::PublishReceived);
-
-        // no stop options, so we just close the socket locally; the broker should send the will
-        stop_client(&will_client).await?;
-
-        let publish_received_events = publish_received_waiter.wait().await?;
-        assert_eq!(1, publish_received_events.len());
-        let publish_received_event = &publish_received_events[0].event;
-        assert_matches!(**publish_received_event, ClientEvent::PublishReceived(_));
-        if let ClientEvent::PublishReceived(event) = &**publish_received_event {
-            verify_publish_received(event, &will_topic, QualityOfService::AtLeastOnce, payload.as_slice());
-        } else {
-            panic!("impossible");
-        }
-
-        stop_client(&client).await?;
-
-        Ok(())
+        async_will_test(builder, build_tokio_client, TokioClientEventWaiter::new_single).await
     }
 
     #[test]
@@ -1516,30 +1285,7 @@ pub(crate) mod testing {
 
     async fn tokio_connect_disconnect_cycle_session_rejoin_test(builder: GenericClientBuilder) -> MqttResult<()> {
         let client = builder.build_tokio(&tokio::runtime::Handle::current()).unwrap();
-        start_client(&client).await?;
-        stop_client(&client).await?;
-
-        for _ in 0..5 {
-            let waiter_config = ClientEventWaiterOptions {
-                wait_type: ClientEventWaitType::Predicate(Box::new(|ev| {
-                    if let ClientEvent::ConnectionSuccess(success_event) = &**ev {
-                        return success_event.connack.session_present && success_event.settings.rejoined_session;
-                    }
-
-                    false
-                })),
-            };
-            let mut connection_success_waiter = TokioClientEventWaiter::new(client.clone(), waiter_config, 1);
-
-            client.start(None)?;
-
-            let connection_success_events = connection_success_waiter.wait().await?;
-            assert_eq!(1, connection_success_events.len());
-
-            stop_client(&client).await?;
-        }
-
-        Ok(())
+        async_connect_disconnect_cycle_session_rejoin_test(client, TokioClientEventWaiter::new_single, TokioClientEventWaiter::new).await
     }
 
     #[test]
